feat(simd_nightly): 30-type portable-simd backend (round-3 fleet)
Complete the portable-simd backend started in the scaffold commit.
12 Sonnet agents (round-3-portable-simd fleet) populated each of the
12 sub-files in `src/simd_nightly/` via the A2A blackboard pattern at
`.claude/board/AGENT_LOG.md`.

Total: ~4,022 LOC of wrapper code + 76 parity tests.

Per-file (line counts at commit):
  - f32_types.rs (395)    — F32x16, F32x8
  - f64_types.rs (307)    — F64x8, F64x4
  - u8_types.rs (1043)    — U8x32, U8x64 + 26 in-file tests
  - u_word_types.rs (520) — U16x32, U32x16, U32x8, U64x8, U64x4
  - i8_types.rs (263)     — I8x32, I8x64
  - i_word_types.rs (449) — I16x16, I16x32, I32x16, I64x8
  - masks.rs (196)        — F32Mask16, F32Mask8, F64Mask8, F64Mask4
  - bf16_types.rs (248)   — BF16x16, BF16x8 (scalar emulation;
                            core::simd has no half-precision)
  - f16_types.rs (220)    — F16x16 (scalar IEEE-754 binary16 emulation)
  - ops.rs (265)          — Add/Sub/Mul/Div/Neg + bitwise + Default
                            macros, applied to all 17 numeric types
  - exotic_methods.rs (329) — permute_bytes / shuffle_bytes / mask_blend /
                              unpack_lo_epi8 / unpack_hi_epi8 scalar
                              fallbacks for U8x32 + U8x64 (core::simd
                              has no native cross-lane byte ops or
                              bitmask-driven blend)
  - tests.rs (815)        — 76 parity tests vs scalar reference

30 types total (mirrors the AVX-512 / AVX2 polyfill surface 1:1).
All re-exported flat from `crate::simd_nightly::*` via the mod.rs
aggregator.

Verification:
  rustup run nightly cargo check --features nightly-simd -p ndarray --lib
    → Finished, 0 errors
  rustup run nightly cargo test --features nightly-simd -p ndarray --lib simd_nightly
    → test result: ok. 153 passed; 0 failed
  cargo check --lib (stable, default features, no nightly-simd)
    → Finished, 0 errors (the existing intrinsics dispatch is unchanged)

Cross-agent findings worth folding into a handover note:
  - `std::simd::StdFloat` is the trait that provides mul_add/sqrt/round/
    floor on core::simd float vectors. `core::simd::num::SimdFloat`
    provides reduce/min/max/clamp but NOT the transcendentals.
  - `core::simd::cmp::SimdOrd` is needed for simd_min/simd_max on
    integer vectors (SimdPartialOrd alone is not sufficient).
  - `core::simd::Mask::to_bitmask()` always returns u64 regardless of
    lane count. Wrappers cast `as u8` / `as u16` / `as u32` for narrower
    bitmask shapes.
  - `core::simd::Simd::swizzle` is `const N: usize` — cannot take a
    runtime index vector. permute_bytes / shuffle_bytes need scalar
    fallback. Same shape as the AVX-512F-without-VBMI fallback path in
    simd_avx512.rs added in PR #142.

What this enables:
  Miri can execute every method here (intrinsics-based backends are
  opaque to miri). Consumers who want miri-runnable SIMD tests import
  from `ndarray::simd_nightly::*` explicitly. The main polyfill via
  `crate::simd::*` continues to use intrinsics — the nightly-simd
  feature does NOT replace the production dispatch, it provides a
  parallel namespace for miri tooling.

Fleet output in .claude/board/AGENT_LOG.md (round-3-portable-simd
section). 6 of 12 agents hit the same AGENT_LOG-write permission
pre-existing block from round-2 — backfilled by the main thread.

+++ src/simd_nightly/i_word_types.rs
@@ -2,7 +2,7 @@
 #![cfg(feature = "nightly-simd")]
 
 use core::simd::{i16x16, i16x32, i32x16, i64x8};
-use core::simd::cmp::{SimdPartialEq, SimdPartialOrd};
+use core::simd::cmp::{SimdOrd, SimdPartialEq, SimdPartialOrd};
 use core::simd::num::SimdInt;
 
 // ════════════════════════════════════════════════════════════════════
@@ -94,13 +94,13 @@ impl I16x16 {
     /// Returns a `u16` matching the 16-lane count.
     #[inline(always)]
     pub fn cmpeq_mask(self, other: Self) -> u16 {
-        self.0.simd_eq(other.0).to_bitmask()
+        self.0.simd_eq(other.0).to_bitmask() as u16
     }
 
     /// Per-lane signed greater-than. Bit i set iff `self[i] > other[i]`.
     #[inline(always)]
     pub fn cmpgt_mask(self, other: Self) -> u16 {
-        self.0.simd_gt(other.0).to_bitmask()
+        self.0.simd_gt(other.0).to_bitmask() as u16
     }
 }
 
@@ -206,13 +206,13 @@ impl I16x32 {
     /// Returns a `u32` matching the 32-lane count.
     #[inline(always)]
     pub fn cmpeq_mask(self, other: Self) -> u32 {
-        self.0.simd_eq(other.0).to_bitmask()
+        self.0.simd_eq(other.0).to_bitmask() as u32
     }
 
     /// Per-lane signed greater-than. Bit i set iff `self[i] > other[i]`.
     #[inline(always)]
     pub fn cmpgt_mask(self, other: Self) -> u32 {
-        self.0.simd_gt(other.0).to_bitmask()
+        self.0.simd_gt(other.0).to_bitmask() as u32
     }
 }
 
@@ -306,13 +306,13 @@ impl I32x16 {
     /// Returns a `u16` matching the 16-lane count.
     #[inline(always)]
     pub fn cmpeq_mask(self, other: Self) -> u16 {
-        self.0.simd_eq(other.0).to_bitmask()
+        self.0.simd_eq(other.0).to_bitmask() as u16
     }
 
     /// Per-lane signed greater-than. Bit i set iff `self[i] > other[i]`.
     #[inline(always)]
     pub fn cmpgt_mask(self, other: Self) -> u16 {
-        self.0.simd_gt(other.0).to_bitmask()
+        self.0.simd_gt(other.0).to_bitmask() as u16
     }
 }
 
@@ -406,13 +406,13 @@ impl I64x8 {
     /// Returns a `u8` matching the 8-lane count.
     #[inline(always)]
     pub fn cmpeq_mask(self, other: Self) -> u8 {
-        self.0.simd_eq(other.0).to_bitmask()
+        self.0.simd_eq(other.0).to_bitmask() as u8
     }
 
     /// Per-lane signed greater-than. Bit i set iff `self[i] > other[i]`.
     #[inline(always)]
     pub fn cmpgt_mask(self, other: Self) -> u8 {
-        self.0.simd_gt(other.0).to_bitmask()
+        self.0.simd_gt(other.0).to_bitmask() as u8
     }
 }
